refactor: complete moving to react-native
