+++ android/app/BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.android_ligen",
+    package = "com.android_libgen",
 )
 
 android_resource(
     name = "res",
-    package = "com.android_ligen",
+    package = "com.android_libgen",
     res = "src/main/res",
 )
