+++ android/app/src/main/java/com/android_libgen/MainApplication.java
@@ -1,4 +1,4 @@
-package com.android_ligen;
+package com.android_libgen;
 
 import android.app.Application;
 import android.content.Context;
