+++ android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">android_ligen</string>
+    <string name="app_name">android_libgen</string>
 </resources>
