+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "android_ligen",
-  "displayName": "android_ligen"
+  "name": "android_libgen",
+  "displayName": "Android Libgen"
 }
