+++ ios/android_libgen.xcodeproj/xcshareddata/xcschemes/android_libgen-tvOS.xcscheme
@@ -29,9 +29,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-               BuildableName = "android_ligen-tvOS.app"
-               BlueprintName = "android_ligen-tvOS"
-               ReferencedContainer = "container:android_ligen.xcodeproj">
+               BuildableName = "android_libgen-tvOS.app"
+               BlueprintName = "android_libgen-tvOS"
+               ReferencedContainer = "container:android_libgen.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
          <BuildActionEntry
@@ -43,9 +43,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
-               BuildableName = "android_ligen-tvOSTests.xctest"
-               BlueprintName = "android_ligen-tvOSTests"
-               ReferencedContainer = "container:android_ligen.xcodeproj">
+               BuildableName = "android_libgen-tvOSTests.xctest"
+               BlueprintName = "android_libgen-tvOSTests"
+               ReferencedContainer = "container:android_libgen.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -61,9 +61,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
-               BuildableName = "android_ligen-tvOSTests.xctest"
-               BlueprintName = "android_ligen-tvOSTests"
-               ReferencedContainer = "container:android_ligen.xcodeproj">
+               BuildableName = "android_libgen-tvOSTests.xctest"
+               BlueprintName = "android_libgen-tvOSTests"
+               ReferencedContainer = "container:android_libgen.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -71,9 +71,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "android_ligen-tvOS.app"
-            BlueprintName = "android_ligen-tvOS"
-            ReferencedContainer = "container:android_ligen.xcodeproj">
+            BuildableName = "android_libgen-tvOS.app"
+            BlueprintName = "android_libgen-tvOS"
+            ReferencedContainer = "container:android_libgen.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
       <AdditionalOptions>
@@ -94,9 +94,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "android_ligen-tvOS.app"
-            BlueprintName = "android_ligen-tvOS"
-            ReferencedContainer = "container:android_ligen.xcodeproj">
+            BuildableName = "android_libgen-tvOS.app"
+            BlueprintName = "android_libgen-tvOS"
+            ReferencedContainer = "container:android_libgen.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
       <AdditionalOptions>
@@ -113,9 +113,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "android_ligen-tvOS.app"
-            BlueprintName = "android_ligen-tvOS"
-            ReferencedContainer = "container:android_ligen.xcodeproj">
+            BuildableName = "android_libgen-tvOS.app"
+            BlueprintName = "android_libgen-tvOS"
+            ReferencedContainer = "container:android_libgen.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
