+++ ios/android_libgen/Images.xcassets/AppIcon.appiconset/Contents.json
