+++ ios/android_libgenTests/android_libgenTests.m
@@ -14,11 +14,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React"
 
-@interface android_ligenTests : XCTestCase
+@interface android_libgenTests : XCTestCase
 
 @end
 
-@implementation android_ligenTests
+@implementation android_libgenTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
